test: move tests
